Change timeout to be time between received messages.
The timeout was previously 15 seconds from when the  earliest message
was sent.  This change starts the 15 second timer when the first message
is sent and resets it to another 15 seconds whenever a response message
is received and there are still pending messages.  If a response is
received and there are no more pending messages, the timer is stopped.

This is to address slow disks not working for #2206 but does not improve
the performance slow disks.

Signed-off-by: Keith Lucas <keith.lucas@suse.com>